pyproject.toml
[tool.poetry]
name = "argo"
version = "0.1.0"
description = ""
authors = ["AI3Labs2049 <ai3labs@peopleai.network>"]
readme = "README.md"


packages = [
    { include = "argo", from = "src" }
]


classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = ">=3.10.10,<4.0"
langchain-community="^0.3.3"
langchain-openai="^0.2.14"
langgraph="^0.2.39"
fastapi = "^0.115.6"
redis = "^5.2.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
python-decouple = "^3.8"
aiofiles = "^24.1.0"
websockets = "^14.1"
aiohttp = "^3.11.11"
python-multipart = "^0.0.20"
aioconsole = "^0.8.1"
jsonschema = "^4.23.0"
langchain-anthropic = "^0.3.1"
tavily-python = "^0.5.0"
beanie = "^1.29.0"
python-ulid = "^3.0.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
